Move scanning baseline out of MainActivity, so it can be shared with others
