+++ pretixscan/app/src/main/java/eu/pretix/pretixscan/droid/ui/MainActivity.kt
@@ -4,29 +4,20 @@ import android.Manifest
 import android.animation.LayoutTransition
 import android.animation.ObjectAnimator
 import android.animation.ValueAnimator
-import android.app.ProgressDialog
 import android.content.Context
 import android.content.DialogInterface
 import android.content.Intent
-import android.content.RestrictionsManager
 import android.content.pm.PackageManager
 import android.content.res.ColorStateList
 import android.graphics.drawable.Drawable
-import android.media.AudioManager
 import android.media.MediaPlayer
-import android.net.ConnectivityManager
-import android.net.Network
-import android.net.NetworkCapabilities
-import android.net.NetworkRequest
 import android.os.Build
 import android.os.Bundle
 import android.os.Handler
 import android.os.ResultReceiver
 import android.text.SpannableString
 import android.text.SpannableStringBuilder
-import android.util.Base64
 import android.util.DisplayMetrics
-import android.view.KeyEvent
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
@@ -35,10 +26,7 @@ import android.view.animation.Animation
 import android.view.animation.BounceInterpolator
 import android.view.animation.DecelerateInterpolator
 import android.view.animation.LinearInterpolator
-import android.widget.TextView
 import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import androidx.appcompat.widget.AppCompatButton
 import androidx.appcompat.widget.SearchView
 import androidx.core.animation.doOnEnd
 import androidx.core.app.ActivityOptionsCompat
@@ -49,18 +37,14 @@ import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.updatePadding
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ObservableField
-import androidx.preference.PreferenceManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.vectordrawable.graphics.drawable.Animatable2Compat
 import androidx.vectordrawable.graphics.drawable.AnimatedVectorDrawableCompat
 import com.fasterxml.jackson.databind.DeserializationFeature
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.module.SimpleModule
-import com.google.android.material.dialog.MaterialAlertDialogBuilder
-import com.google.android.material.snackbar.Snackbar
 import eu.pretix.libpretixsync.api.PretixApi
 import eu.pretix.libpretixsync.check.CheckException
-import eu.pretix.libpretixsync.check.OnlineCheckProvider
 import eu.pretix.libpretixsync.check.TicketCheckProvider
 import eu.pretix.libpretixsync.db.Answer
 import eu.pretix.libpretixsync.models.db.toModel
@@ -74,46 +58,29 @@ import eu.pretix.libpretixui.android.questions.QuestionsDialogInterface
 import eu.pretix.libpretixui.android.scanning.HardwareScanner
 import eu.pretix.libpretixui.android.scanning.ScanReceiver
 import eu.pretix.libpretixui.android.scanning.ScannerView
-import eu.pretix.pretixscan.droid.*
-import eu.pretix.pretixscan.droid.connectivity.ConnectivityChangedListener
+import eu.pretix.pretixscan.droid.AndroidHttpClientFactory
+import eu.pretix.pretixscan.droid.BuildConfig
+import eu.pretix.pretixscan.droid.PretixScan
+import eu.pretix.pretixscan.droid.R
 import eu.pretix.pretixscan.droid.databinding.ActivityMainBinding
 import eu.pretix.pretixscan.droid.hardware.LED
-import eu.pretix.pretixscan.droid.ui.ResultState.*
+import eu.pretix.pretixscan.droid.ui.ResultState.DIALOG
+import eu.pretix.pretixscan.droid.ui.ResultState.EMPTY
+import eu.pretix.pretixscan.droid.ui.ResultState.ERROR
+import eu.pretix.pretixscan.droid.ui.ResultState.LOADING
+import eu.pretix.pretixscan.droid.ui.ResultState.SUCCESS
+import eu.pretix.pretixscan.droid.ui.ResultState.SUCCESS_EXIT
+import eu.pretix.pretixscan.droid.ui.ResultState.WARNING
 import eu.pretix.pretixscan.droid.ui.info.EventinfoActivity
 import io.sentry.Sentry
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import org.json.JSONArray
 import org.json.JSONObject
 import splitties.toast.toast
-import java.io.IOException
-import java.lang.Integer.max
-import java.nio.charset.Charset
-import java.security.MessageDigest
 import java.text.SimpleDateFormat
 import java.time.format.DateTimeFormatter
-import java.util.*
 
 
-interface ReloadableActivity {
-    fun reload()
-}
-
-enum class ResultCardState {
-    HIDDEN,
-    SHOWN
-}
-
-enum class ResultState {
-    EMPTY,
-    LOADING,
-    ERROR,
-    DIALOG,
-    WARNING,
-    SUCCESS,
-    SUCCESS_EXIT
-}
 
 class ViewDataHolder(private val ctx: Context) {
     val resultState = ObservableField<ResultState>()
@@ -151,66 +118,28 @@ class ViewDataHolder(private val ctx: Context) {
     }
 
     fun setLed(context: Context, state: ResultState, needsAttention: Boolean) {
-        val led = LED(context)
-        when (state) {
-            EMPTY -> led.off()
-            LOADING -> led.progress()
-            DIALOG, WARNING -> led.attention(blink = needsAttention)
-            ERROR -> led.error()
-            SUCCESS, SUCCESS_EXIT -> led.success(blink = needsAttention)
-        }
+        // FIXME
     }
 }
 
-class MainActivity : AppCompatActivity(), ReloadableActivity, ScannerView.ResultHandler, MediaPlayer.OnCompletionListener, ConnectivityChangedListener {
+class MainActivity : BaseScanActivity() {
 
     private val REQ_EVENT = 1
 
     private lateinit var binding: ActivityMainBinding
-    private lateinit var sm: SyncManager
-    private lateinit var conf: AppConfig
-    private val bgScope = CoroutineScope(Dispatchers.IO)
-    private val handler = Handler()
     private val hideHandler = Handler()
     private var hideAnimation: ValueAnimator? = null
     private var card_state = ResultCardState.HIDDEN
     private var view_data = ViewDataHolder(this)
-    private var mediaPlayers: MutableMap<Int, MediaPlayer> = mutableMapOf()
-
-    private var lastScanTime: Long = 0
-    private var lastScanCode: String = ""
-    private var lastIgnoreUnpaid: Boolean = false
-    private var lastScanResult: TicketCheckProvider.CheckResult? = null
-    private var keyboardBuffer: String = ""
-    private var dialog: QuestionsDialogInterface? = null
-    private var pdialog: ProgressDialog? = null
-    private var networkCallback: ConnectivityManager.NetworkCallback? = null
-
-    private val dataWedgeHelper = DataWedgeHelper(this)
 
     private var searchAdapter: SearchListAdapter? = null
     private var searchFilter = ""
 
     private var syncMessage = ""
 
-    private var pendingPinAction: ((pin: String) -> Unit)? = null
-
     companion object {
         const val PERMISSIONS_REQUEST_CAMERA = 1337
     }
-
-    private val hardwareScanner = HardwareScanner(object : ScanReceiver {
-        override fun scanResult(result: String) {
-            if (result == lastScanCode && System.currentTimeMillis() - lastScanTime < 2500) {
-                return
-            }
-            lastScanTime = System.currentTimeMillis()
-            lastScanCode = result
-            lastIgnoreUnpaid = false
-            lastScanResult = null
-            handleScan(result, null, !conf.unpaidAsk)
-        }
-    })
 
     override fun reload() {
         reloadSyncStatus()
@@ -323,7 +252,7 @@ class MainActivity : AppCompatActivity(), ReloadableActivity, ScannerView.Result
         }
     }
 
-    fun reloadSyncStatus() {
+    override fun reloadSyncStatus() {
         if (conf.lastFailedSync > conf.lastSync || System.currentTimeMillis() - conf.lastDownload > 5 * 60 * 1000) {
             binding.textViewStatus.setTextColor(ContextCompat.getColor(this, R.color.pretix_brand_red))
         } else {
@@ -423,35 +352,11 @@ class MainActivity : AppCompatActivity(), ReloadableActivity, ScannerView.Result
         p0?.seekTo(0)
     }
 
-    @SuppressWarnings("ResourceType")
-    private fun buildMediaPlayer() {
-        val resourceIds = listOf(R.raw.enter, R.raw.exit, R.raw.error, R.raw.beep, R.raw.attention)
-        for (r in resourceIds) {
-            val mediaPlayer = MediaPlayer()
-            mediaPlayer.setAudioStreamType(AudioManager.STREAM_MUSIC)
-            mediaPlayer.setOnCompletionListener(this)
-            // mediaPlayer.setOnErrorListener(this)
-            try {
-                val file = resources.openRawResourceFd(r)
-                try {
-                    mediaPlayer.setDataSource(file.fileDescriptor, file.startOffset, file.length)
-                } finally {
-                    file.close();
-                }
-                mediaPlayer.setVolume(0.2f, 0.2f)
-                mediaPlayer.prepare()
-                mediaPlayers[r] = mediaPlayer
-            } catch (ioe: IOException) {
-            }
-        }
-    }
-
-    public override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        conf = AppConfig(this)
-
-        getRestrictions(this)
+    override fun onCreate(savedInstanceState: Bundle?) {
         binding = DataBindingUtil.setContentView(this, R.layout.activity_main)
+
+        super.onCreate(savedInstanceState)
+
         view_data.resultState.set(EMPTY)
         view_data.scanType.set(conf.scanType)
         view_data.hardwareScan.set(!conf.useCamera)
@@ -481,53 +386,16 @@ class MainActivity : AppCompatActivity(), ReloadableActivity, ScannerView.Result
             WindowInsetsCompat.CONSUMED
         }
 
-        volumeControlStream = AudioManager.STREAM_MUSIC
-        buildMediaPlayer()
-
-        if (!conf.deviceRegistered) {
-            registerDevice()
-            return
-        }
-        try {
-            setupApi()
-        } catch (e: IllegalStateException) {
-            // IllegalStateException is thrown by db access -> Migrations.minVersionCallback
-            Sentry.captureException(e)
-            panicPleaseReinstall()
-            return
-        }
         setUpEventListeners()
 
-        if (conf.synchronizedEvents.isEmpty()) {
-            selectEvent()
-        } else if (conf.lastDownload < 1) {
-            syncNow()
-        }
-        scheduleSync()
         checkPermission(Manifest.permission.CAMERA, PERMISSIONS_REQUEST_CAMERA)
 
         hideCard()
         hideSearchCard()
         binding.cardResult.layoutTransition.enableTransitionType(LayoutTransition.CHANGING)
 
-        if (dataWedgeHelper.isInstalled) {
-            try {
-                dataWedgeHelper.install()
-            } catch (e: IOException) {
-                e.printStackTrace()
-            }
-        }
-
         binding.recyclerViewSearch.layoutManager = LinearLayoutManager(this)
         binding.recyclerViewSearch.addItemDecoration(androidx.recyclerview.widget.DividerItemDecoration(binding.recyclerViewSearch.context, androidx.recyclerview.widget.DividerItemDecoration.VERTICAL))
-
-        supportFragmentManager.setFragmentResultListener(PinDialog.RESULT_PIN, this) { _, bundle ->
-            val pin = bundle.getString(PinDialog.RESULT_PIN)
-            if (pin != null && conf.verifyPin(pin)) {
-                (supportFragmentManager.findFragmentByTag(PinDialog.TAG) as? PinDialog)?.dismiss()
-                pendingPinAction?.let { it(pin) }
-            }
-        }
     }
 
     private fun eventButtonText(): String {
@@ -539,39 +407,16 @@ class MainActivity : AppCompatActivity(), ReloadableActivity, ScannerView.Result
         }
     }
 
-    private fun setupApi() {
+    override fun setupApi() {
         val ebt = eventButtonText()
         if (binding.mainToolbar.event != null && binding.mainToolbar.event.text != ebt) {  // can be null if search bar is open
             binding.mainToolbar.event.text = ebt
             (binding.mainToolbar.event.parent as View?)?.forceLayout()
         }
-        val api = PretixApi.fromConfig(conf, AndroidHttpClientFactory(application as PretixScan))
-
-        sm = SyncManager(
-                conf,
-                api,
-                AndroidSentryImplementation(),
-                (application as PretixScan).db,
-                (application as PretixScan).fileStorage,
-                60000L,
-                5 * 60000L,
-                if (conf.syncOrders) SyncManager.Profile.PRETIXSCAN else SyncManager.Profile.PRETIXSCAN_ONLINE,
-                conf.printBadges,
-                BuildConfig.VERSION_CODE,
-                JSONObject(),
-                Build.BRAND,
-                Build.MODEL,
-                (if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) Build.VERSION.BASE_OS else "").ifEmpty { "Android" },
-                Build.VERSION.RELEASE,
-                "pretixSCAN Android",
-                BuildConfig.VERSION_NAME,
-                null,
-                null,
-                (application as PretixScan).connectivityHelper
-        )
+        super.setupApi()
     }
 
-    private fun selectEvent() {
+    override fun selectEvent() {
         val intent = Intent(this, EventConfigActivity::class.java)
         if (binding.mainToolbar.event != null && ViewCompat.isLaidOut(binding.mainToolbar.event)) {
             val options = ActivityOptionsCompat.makeSceneTransitionAnimation(
@@ -580,26 +425,6 @@ class MainActivity : AppCompatActivity(), ReloadableActivity, ScannerView.Result
         } else {
             startWithPIN(intent, "switch_event", REQ_EVENT, null)
         }
-    }
-
-    private fun snackbar(message: String) {
-        Snackbar.make(findViewById(R.id.root_layout), message, Snackbar.LENGTH_LONG).show();
-    }
-
-    private fun snackbar(message: Int) {
-        Snackbar.make(findViewById(R.id.root_layout), message, Snackbar.LENGTH_LONG).show();
-    }
-
-    private fun registerDevice() {
-        val intent = Intent(this, WelcomeActivity::class.java)
-        startActivity(intent)
-        finish()
-    }
-
-    private fun panicPleaseReinstall() {
-        val intent = Intent(this, PleaseReinstallActivity::class.java)
-        startActivity(intent)
-        finish()
     }
 
     private val hideRunnable = Runnable {
@@ -637,157 +462,8 @@ class MainActivity : AppCompatActivity(), ReloadableActivity, ScannerView.Result
         hideAnimation!!.start()
     }
 
-    private val syncRunnable = Runnable {
-        syncMessage = ""
-        val activity = this
-        bgScope.launch {
-            if (!(application as PretixScan).syncLock.tryLock()) {
-                runOnUiThread {
-                    reloadSyncStatus()
-                }
-                scheduleSync()
-                return@launch
-            }
-            val prefs = PreferenceManager.getDefaultSharedPreferences(activity)
-            try {
-                if (prefs.getBoolean("pref_sync_auto", true)) {
-                    val result = sm.sync(false) {
-                        runOnUiThread {
-                            syncMessage = it
-                            reloadSyncStatus()
-                        }
-                    }
-                    if (result.isDataDownloaded) {
-                        runOnUiThread {
-                            reload()
-                        }
-                    }
-                }
-                runOnUiThread {
-                    reloadSyncStatus()
-                    scheduleSync()
-                }
-            } catch (e: SyncManager.EventSwitchRequested) {
-                runOnUiThread {
-                    conf.eventSelection = listOf(EventSelection(
-                            eventSlug = e.eventSlug,
-                            eventName = e.eventName,
-                            subEventId = e.subeventId,
-                            checkInList = e.checkinlistId,
-                            dateFrom = null,
-                            dateTo = null
-                    ))
-                    setupApi()
-                    syncNow()
-                    reload()
-                }
-            } catch (e: Exception) {
-                runOnUiThread {
-                    reload()
-                }
-            } finally {
-                (application as PretixScan).syncLock.unlock()
-                runOnUiThread {
-                    reloadSyncStatus()
-                }
-            }
-        }
-    }
-
-    fun scheduleSync() {
-        handler.removeCallbacks(syncRunnable)
-        handler.postDelayed(syncRunnable, 1000)
-    }
-
-    fun syncNow() {
-        if (isDestroyed) return
-        syncMessage = ""
-        pdialog = ProgressDialog(this).apply {
-            isIndeterminate = true
-            setMessage(getString(R.string.progress_syncing))
-            setTitle(R.string.progress_syncing)
-            setCanceledOnTouchOutside(false)
-            setCancelable(false)
-            show()
-        }
-        val activity = this
-        bgScope.launch {
-            if (!(application as PretixScan).syncLock.tryLock()) {
-                runOnUiThread {
-                    MaterialAlertDialogBuilder(activity).setMessage(R.string.error_sync_in_background).create().show()
-                    pdialog?.dismiss()
-                }
-                return@launch
-            }
-            try {
-                sm.sync(true) { current_action ->
-                    runOnUiThread {
-                        if (isDestroyed) {
-                            return@runOnUiThread
-                        }
-                        reloadSyncStatus()
-                        syncMessage = current_action
-                        pdialog?.setMessage(current_action)
-                    }
-                }
-                runOnUiThread {
-                    if (isDestroyed) {
-                        return@runOnUiThread
-                    }
-                    reload()
-                    pdialog?.dismiss()
-                    if (conf.lastFailedSync > 0) {
-                        MaterialAlertDialogBuilder(activity).setMessage(conf.lastFailedSyncMsg).create().show()
-                    }
-                }
-            } catch (e: SyncManager.EventSwitchRequested) {
-                runOnUiThread {
-                    if (isDestroyed) {
-                        return@runOnUiThread
-                    }
-                    pdialog?.dismiss()
-                    conf.eventSelection = listOf(EventSelection(
-                            eventSlug = e.eventSlug,
-                            eventName = e.eventName,
-                            subEventId = e.subeventId,
-                            checkInList = e.checkinlistId,
-                            dateFrom = null,
-                            dateTo = null
-                    ))
-                    setupApi()
-                    syncNow()
-                    reload()
-                }
-            } catch (e: Exception) {
-                e.printStackTrace()
-                runOnUiThread {
-                    if (isDestroyed) {
-                        return@runOnUiThread
-                    }
-                    Sentry.captureException(e)
-                    pdialog?.dismiss()
-                    MaterialAlertDialogBuilder(activity)
-                        .setMessage(e.message ?: getString(R.string.error_unknown_exception))
-                        .create().show()
-                }
-            } finally {
-                (application as PretixScan).syncLock.unlock()
-            }
-        }
-    }
-
     override fun onResume() {
         super.onResume()
-        try {
-            reload()
-            setupApi()
-        } catch (e: IllegalStateException) {
-            // IllegalStateException is thrown by db access -> Migrations.minVersionCallback
-            Sentry.captureException(e)
-            panicPleaseReinstall()
-            return
-        }
-        getRestrictions(this)
 
         view_data.kioskMode.set(conf.kioskMode)
         if (conf.kioskMode) {
@@ -795,13 +471,13 @@ class MainActivity : AppCompatActivity(), ReloadableActivity, ScannerView.Result
             window.decorView.apply {
                 systemUiVisibility = View.SYSTEM_UI_FLAG_FULLSCREEN
             }
+            return
         } else {
             supportActionBar?.show()
         }
 
         scheduleSync()
 
-        hardwareScanner.start(this)
         LED(this).off()
 
         if (conf.useCamera && ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED) {
@@ -815,30 +491,12 @@ class MainActivity : AppCompatActivity(), ReloadableActivity, ScannerView.Result
         setKioskAnimation()
 
         reloadCameraState()
+    }
 
-        (application as PretixScan).connectivityHelper.addListener(this)
-
-        val networkRequest = NetworkRequest.Builder()
-                .addTransportType(NetworkCapabilities.TRANSPORT_CELLULAR)
-                .addTransportType(NetworkCapabilities.TRANSPORT_WIFI)
-                .addTransportType(NetworkCapabilities.TRANSPORT_ETHERNET)
-                .build()
-        val connectivityManager = getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-        fun check() {
-            (application as PretixScan).connectivityHelper.setHardOffline(connectivityManager.activeNetworkInfo?.isConnectedOrConnecting != true)
-        }
-
-        networkCallback = object : ConnectivityManager.NetworkCallback() {
-            override fun onAvailable(network: Network) {
-                (application as PretixScan).connectivityHelper.setHardOffline(false)
-            }
-
-            override fun onLost(network: Network) {
-                check()
-            }
-        }
-        check()
-        connectivityManager.registerNetworkCallback(networkRequest, networkCallback!!)
+    fun launchKiosk() {
+        val intent = Intent(this, KioskActivity::class.java)
+        startActivity(intent)
+        finish()
     }
 
     private fun setKioskAnimation() {
@@ -964,163 +622,32 @@ class MainActivity : AppCompatActivity(), ReloadableActivity, ScannerView.Result
     }
 
     override fun onPause() {
-        handler.removeCallbacks(syncRunnable)
-        (application as PretixScan).connectivityHelper.removeListener(this)
         super.onPause()
         if (conf.useCamera) {
             binding.scannerView.stopCamera()
         }
-        hardwareScanner.stop(this)
         LED(this).off()
-
-        if (networkCallback != null) {
-            val connectivityManager = getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-            connectivityManager.unregisterNetworkCallback(networkCallback!!)
-        }
     }
 
-    override fun onStop() {
-        LED(this).off()
-        super.onStop()
-    }
-
-    fun handleScan(raw_result: String, answers: MutableList<Answer>?, ignore_unpaid: Boolean = false) {
-        if (conf.kioskMode && conf.requiresPin("settings") && conf.verifyPin(raw_result)) {
-            supportActionBar?.show()
-            return
-        }
-
+    override fun handleScan(raw_result: String, answers: MutableList<Answer>?, ignore_unpaid: Boolean) {
         if (dialog?.isShowing() == true) {
             /*
              * Skip scan if a dialog is still in front. This forces users to answer the questions asked.
              */
             return
         }
+        LED(this).off()
 
-        val result = if (Regex("^HC1:[0-9A-Z $%*+-./:]+$").matches(raw_result.uppercase(Locale.getDefault()))) {
-            /*
-             * This is a bit of a hack. pretixSCAN 1.11-2.8.2 supports checking digital COVID vaccination
-             * certificates. When scanning them at the correct time, we have a high level of privacy
-             * since we do not store any personal data contained in the certificate. However, if you
-             * accidentally scan the certificate when you are supposed to scan a ticket, our fancy
-             * error log will cause the verbatim vaccination certificate to be stored on the server.
-             * Not really our fault, but also not really nice to store that sensitive health info.
-             * However, it's still helpful for debugging to see how often an invalid code was scanned.
-             * So if we encounter something that looks like an EU DGC, we'll just transform it into
-             * a hashed version.
-             *
-             * This hack is safe for pretix' default signature schemes, as they would never generate
-             * a QR code starting with ``HC1:``, but it could theoretically be unsafe for third-party
-             * plugins.
-             */
-            val md = MessageDigest.getInstance("SHA-256")
-            md.update(raw_result.toByteArray(Charset.defaultCharset()))
-            "HC1:hashed:" + Base64.encodeToString(md.digest(), Base64.URL_SAFE)
-        } else {
-            raw_result
+        if (conf.kioskMode && conf.requiresPin("settings") && conf.verifyPin(raw_result)) {
+            supportActionBar?.show()
+            return
         }
-
         showLoadingCard()
         hideSearchCard()
-
-        if (answers == null && !ignore_unpaid && !conf.offlineMode && conf.sounds) {
-            mediaPlayers[R.raw.beep]?.start()
-        }
-
-        bgScope.launch {
-            var checkResult: TicketCheckProvider.CheckResult? = null
-            val provider = (application as PretixScan).getCheckProvider(conf)
-            val startedAt = System.currentTimeMillis()
-            try {
-                checkResult = provider.check(conf.eventSelectionToMap(), result, "barcode", answers, ignore_unpaid, conf.printBadges, when (conf.scanType) {
-                    "exit" -> TicketCheckProvider.CheckInType.EXIT
-                    else -> TicketCheckProvider.CheckInType.ENTRY
-                }, allowQuestions = !conf.ignoreQuestions)
-                if (provider is OnlineCheckProvider) {
-                    if (checkResult?.type == TicketCheckProvider.CheckResult.Type.ERROR) {
-                        (application as PretixScan).connectivityHelper.recordError()
-                    } else {
-                        (application as PretixScan).connectivityHelper.recordSuccess(System.currentTimeMillis() - startedAt)
-                    }
-                }
-            } catch (e: Exception) {
-                Sentry.captureException(e)
-                if (BuildConfig.DEBUG) {
-                    e.printStackTrace()
-                }
-                (application as PretixScan).connectivityHelper.recordError()
-                checkResult = TicketCheckProvider.CheckResult(TicketCheckProvider.CheckResult.Type.INVALID, getString(R.string.error_unknown_exception))
-            }
-            runOnUiThread {
-                displayScanResult(checkResult!!, answers, ignore_unpaid)
-            }
-        }
+        super.handleScan(raw_result, answers, ignore_unpaid)
     }
 
-    fun showQuestionsDialog(res: TicketCheckProvider.CheckResult, secret: String, ignore_unpaid: Boolean,
-                            values: Map<String, String>?, isResumed: Boolean,
-                            retryHandler: ((String, MutableList<Answer>, Boolean) -> Unit)): QuestionsDialogInterface {
-
-        val questions = res.requiredAnswers!!.map { it.question.toModel() }
-        for (q in questions) {
-            q.resolveDependency(questions)
-        }
-
-        val values_ = if (values == null) {
-            val v = mutableMapOf<String, String>()
-            res.requiredAnswers!!.forEach {
-                if (!it.currentValue.isNullOrBlank()) {
-                    v[it.question.toModel().identifier] = it.currentValue!!
-                }
-            }
-            v
-        } else {
-            values
-        }
-        val attendeeName = if (conf.hideNames || res.position?.isNull("attendee_name") != false) {
-            ""
-        } else {
-            res.position?.optString("attendee_name")
-        }
-
-        var attendeeDOB: String? = null
-        if (!conf.hideNames) {
-            val qlen = res.position?.getJSONArray("answers")?.length() ?: 0
-            for (i in 0 until qlen) {
-                val answ = res.position!!.getJSONArray("answers")!!.getJSONObject(i)
-                if (answ.getString("question_identifier") == "dob") {
-                    attendeeDOB = answ.getString("answer")
-                }
-            }
-        }
-
-        return eu.pretix.libpretixui.android.questions.showQuestionsDialog(
-                QuestionsDialog.Companion.QuestionsType.LINE_ITEM_QUESTIONS,
-                this,
-                questions,
-                values_,
-                null,
-                null,
-                { answers -> retryHandler(secret, answers, ignore_unpaid) },
-                null,
-                attendeeName,
-                attendeeDOB,
-                res.orderCodeAndPositionId(),
-                if (res.ticket != null) {
-                    if (res.variation != null) {
-                        res.ticket + " – " + res.variation
-                    } else {
-                        res.ticket
-                    }
-                } else {
-                    null
-                },
-                !conf.useCamera,
-                isResumed
-        )
-    }
-
-    fun displayScanResult(result: TicketCheckProvider.CheckResult, answers: MutableList<Answer>?, ignore_unpaid: Boolean = false) {
+    override fun displayScanResult(result: TicketCheckProvider.CheckResult, answers: MutableList<Answer>?, ignore_unpaid: Boolean) {
         lastScanResult = result
         lastIgnoreUnpaid = ignore_unpaid
 
@@ -1364,63 +891,7 @@ class MainActivity : AppCompatActivity(), ReloadableActivity, ScannerView.Result
         if ((dialog != null && dialog!!.isShowing()) || view_data.resultState.get() == LOADING) {
             return
         }
-
-        val s = rawResult.text
-        if (s == lastScanCode && System.currentTimeMillis() - lastScanTime < 5000) {
-            return
-        }
-        lastScanTime = System.currentTimeMillis()
-        lastScanCode = s
-        lastScanResult = null
-        lastIgnoreUnpaid = false
-        handleScan(s, null, !conf.unpaidAsk)
-    }
-
-    override fun dispatchKeyEvent(event: KeyEvent): Boolean {
-        if ((event.action != KeyEvent.ACTION_DOWN && event.action != KeyEvent.ACTION_MULTIPLE) || (currentFocus is TextView && currentFocus !is AppCompatButton)) {
-            return super.dispatchKeyEvent(event)
-        }
-        return when (event.keyCode) {
-            KeyEvent.KEYCODE_ENTER -> {
-                if (keyboardBuffer.isEmpty()) {
-                    false
-                }
-                lastScanTime = System.currentTimeMillis()
-                lastScanCode = keyboardBuffer
-                lastScanResult = null
-                lastIgnoreUnpaid = false
-                handleScan(keyboardBuffer, null, !conf.unpaidAsk)
-                keyboardBuffer = ""
-                true
-            }
-            KeyEvent.KEYCODE_UNKNOWN -> {
-                keyboardBuffer += event.characters
-                true
-            }
-            else -> {
-                val codepoint = event.keyCharacterMap.get(event.keyCode, event.metaState)
-                if (codepoint > 0) {
-                    keyboardBuffer += codepoint.toChar().toString().repeat(max(event.repeatCount, 1))
-                    true
-                } else {
-                    super.dispatchKeyEvent(event)
-                }
-            }
-        }
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        if (requestCode == REQ_EVENT) {
-            if (resultCode == RESULT_OK) {
-                setupApi()
-                scheduleSync()
-                reload()
-            }
-        } else if (dialog?.handleActivityResult(requestCode, resultCode, data) == true) {
-            return
-        } else {
-            super.onActivityResult(requestCode, resultCode, data)
-        }
+        super.handleResult(rawResult)
     }
 
 
@@ -1467,26 +938,6 @@ class MainActivity : AppCompatActivity(), ReloadableActivity, ScannerView.Result
         return super.onCreateOptionsMenu(menu)
     }
 
-    fun pinProtect(key: String, valid: ((pin: String) -> Unit)) {
-        if (!conf.requiresPin(key)) {
-            valid("")
-            return
-        }
-        pendingPinAction = valid
-        PinDialog().show(supportFragmentManager)
-    }
-
-    fun startWithPIN(intent: Intent, key: String, resultCode: Int? = null, bundle: Bundle? = null) {
-        pinProtect(key) { pin ->
-            intent.putExtra("pin", pin)
-            if (resultCode != null) {
-                startActivityForResult(intent, resultCode, bundle)
-            } else {
-                startActivity(intent)
-            }
-        }
-    }
-
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         when (item.itemId) {
             R.id.action_settings -> {
@@ -1516,28 +967,6 @@ class MainActivity : AppCompatActivity(), ReloadableActivity, ScannerView.Result
             }
         }
         return super.onOptionsItemSelected(item)
-    }
-
-
-    fun getRestrictions(ctx: Context) {
-        val myRestrictionsMgr = ctx.getSystemService(Context.RESTRICTIONS_SERVICE) as RestrictionsManager?
-                ?: return
-        val restrictions = myRestrictionsMgr.applicationRestrictions
-        val prefs = PreferenceManager.getDefaultSharedPreferences(ctx)
-
-        for (key in restrictions.keySet()) {
-            if (key == "pref_auto_print_badges_option") {
-                prefs.edit().putString(key, restrictions.getString(key)).apply()
-            } else {
-                prefs.edit().putBoolean(key, restrictions.getBoolean(key)).apply()
-            }
-        }
-    }
-
-    override fun onConnectivityChanged() {
-        runOnUiThread {
-            reload()
-        }
     }
 
     override fun onRestoreInstanceState(savedInstanceState: Bundle) {
